Fix issue with coordinates

+++ photobox/models/Area.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass
 @dataclass_json(letter_case=LetterCase.CAMEL, undefined=Undefined.EXCLUDE)
 @dataclass
 class Area:
-    x: int
-    y: int
-    width: int
-    height: int
+    x: float
+    y: float
+    width: float
+    height: float
